♻️ refactor: SetUserProfile 컴포넌트 폼 입력 부분 분리 (#78)
